--- Assets/Scripts/Procedural Terrain/Data/Updatable Data.cs
@@ -1,26 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class UpdatableData : ScriptableObject {
-
-    public event System.Action onValuesUpdated;
-    public bool autoUpdate;
-
-    public void notifyOfUpdatedValues() {
-
-        UnityEditor.EditorApplication.update -= notifyOfUpdatedValues;
-
-        if(onValuesUpdated != null) {
-            onValuesUpdated();
-        }
-
-    }
-
-    protected virtual void OnValidate() {
-        if(autoUpdate) {
-            UnityEditor.EditorApplication.update += notifyOfUpdatedValues;
-        }
-    }
-
-}
