+++ Assets/Standard Assets/Characters/FirstPersonCharacter/Scripts/RigidbodyFirstPersonController.cs
@@ -1,54 +1,63 @@
 using System;
 using UnityEngine;
 
-namespace UnityStandardAssets.Characters.FirstPerson {
-    [RequireComponent(typeof(Rigidbody))]
-    [RequireComponent(typeof(CapsuleCollider))]
-    public class RigidbodyFirstPersonController : MonoBehaviour {
+namespace UnityStandardAssets.Characters.FirstPerson
+{
+    [RequireComponent(typeof (Rigidbody))]
+    [RequireComponent(typeof (CapsuleCollider))]
+    public class RigidbodyFirstPersonController : MonoBehaviour
+    {
         [Serializable]
-        public class MovementSettings {
+        public class MovementSettings
+        {
             public float ForwardSpeed = 8.0f;   // Speed when walking forward
             public float BackwardSpeed = 4.0f;  // Speed when walking backwards
             public float StrafeSpeed = 4.0f;    // Speed when walking sideways
             public float RunMultiplier = 2.0f;   // Speed when sprinting
-            public KeyCode RunKey = KeyCode.LeftShift;
+	        public KeyCode RunKey = KeyCode.LeftShift;
             public float JumpForce = 30f;
             public AnimationCurve SlopeCurveModifier = new AnimationCurve(new Keyframe(-90.0f, 1.0f), new Keyframe(0.0f, 1.0f), new Keyframe(90.0f, 0.0f));
-            [HideInInspector]
-            public float CurrentTargetSpeed = 8f;
+            [HideInInspector] public float CurrentTargetSpeed = 8f;
 
 #if !MOBILE_INPUT
             private bool m_Running;
 #endif
 
-            public void UpdateDesiredTargetSpeed(Vector2 input) {
-                if(input == Vector2.zero)
-                    return;
-                if(input.x > 0 || input.x < 0) {
-                    //strafe
-                    CurrentTargetSpeed = StrafeSpeed;
-                }
-                if(input.y < 0) {
-                    //backwards
-                    CurrentTargetSpeed = BackwardSpeed;
-                }
-                if(input.y > 0) {
-                    //forwards
-                    //handled last as if strafing and moving forward at the same time forwards speed should take precedence
-                    CurrentTargetSpeed = ForwardSpeed;
-                }
+            public void UpdateDesiredTargetSpeed(Vector2 input)
+            {
+	            if (input == Vector2.zero) return;
+				if (input.x > 0 || input.x < 0)
+				{
+					//strafe
+					CurrentTargetSpeed = StrafeSpeed;
+				}
+				if (input.y < 0)
+				{
+					//backwards
+					CurrentTargetSpeed = BackwardSpeed;
+				}
+				if (input.y > 0)
+				{
+					//forwards
+					//handled last as if strafing and moving forward at the same time forwards speed should take precedence
+					CurrentTargetSpeed = ForwardSpeed;
+				}
 #if !MOBILE_INPUT
-                if(Input.GetKey(RunKey)) {
-                    CurrentTargetSpeed *= RunMultiplier;
-                    m_Running = true;
-                } else {
-                    m_Running = false;
-                }
+	            if (Input.GetKey(RunKey))
+	            {
+		            CurrentTargetSpeed *= RunMultiplier;
+		            m_Running = true;
+	            }
+	            else
+	            {
+		            m_Running = false;
+	            }
 #endif
             }
 
 #if !MOBILE_INPUT
-            public bool Running {
+            public bool Running
+            {
                 get { return m_Running; }
             }
 #endif
@@ -56,7 +65,8 @@ namespace UnityStandardAssets.Characters.FirstPerson {
 
 
         [Serializable]
-        public class AdvancedSettings {
+        public class AdvancedSettings
+        {
             public float groundCheckDistance = 0.01f; // distance for checking if the controller is grounded ( 0.01f seems to work best for this )
             public float stickToGroundHelperDistance = 0.5f; // stops the character
             public float slowDownRate = 20f; // rate at which the controller comes to a stop when there is no input
@@ -79,22 +89,27 @@ namespace UnityStandardAssets.Characters.FirstPerson {
         private bool m_Jump, m_PreviouslyGrounded, m_Jumping, m_IsGrounded;
 
 
-        public Vector3 Velocity {
+        public Vector3 Velocity
+        {
             get { return m_RigidBody.velocity; }
         }
 
-        public bool Grounded {
+        public bool Grounded
+        {
             get { return m_IsGrounded; }
         }
 
-        public bool Jumping {
+        public bool Jumping
+        {
             get { return m_Jumping; }
         }
 
-        public bool Running {
-            get {
-#if !MOBILE_INPUT
-                return movementSettings.Running;
+        public bool Running
+        {
+            get
+            {
+ #if !MOBILE_INPUT
+				return movementSettings.Running;
 #else
 	            return false;
 #endif
@@ -102,57 +117,68 @@ namespace UnityStandardAssets.Characters.FirstPerson {
         }
 
 
-        private void Start() {
+        private void Start()
+        {
             m_RigidBody = GetComponent<Rigidbody>();
             m_Capsule = GetComponent<CapsuleCollider>();
-            mouseLook.Init(transform, cam.transform);
+            mouseLook.Init (transform, cam.transform);
         }
 
 
-        private void Update() {
-
+        private void Update()
+        {
             RotateView();
 
-            if(Input.GetButtonDown("Jump") && !m_Jump) {
+            if (Input.GetButtonDown("Jump") && !m_Jump)
+            {
                 m_Jump = true;
             }
         }
 
 
-        private void FixedUpdate() {
+        private void FixedUpdate()
+        {
             GroundCheck();
             Vector2 input = GetInput();
 
-            if((Mathf.Abs(input.x) > float.Epsilon || Mathf.Abs(input.y) > float.Epsilon) && (advancedSettings.airControl || m_IsGrounded)) {
+            if ((Mathf.Abs(input.x) > float.Epsilon || Mathf.Abs(input.y) > float.Epsilon) && (advancedSettings.airControl || m_IsGrounded))
+            {
                 // always move along the camera forward as it is the direction that it being aimed at
-                Vector3 desiredMove = cam.transform.forward * input.y + cam.transform.right * input.x;
+                Vector3 desiredMove = cam.transform.forward*input.y + cam.transform.right*input.x;
                 desiredMove = Vector3.ProjectOnPlane(desiredMove, m_GroundContactNormal).normalized;
 
-                desiredMove.x = desiredMove.x * movementSettings.CurrentTargetSpeed;
-                desiredMove.z = desiredMove.z * movementSettings.CurrentTargetSpeed;
-                desiredMove.y = desiredMove.y * movementSettings.CurrentTargetSpeed;
-                if(m_RigidBody.velocity.sqrMagnitude <
-                    (movementSettings.CurrentTargetSpeed * movementSettings.CurrentTargetSpeed)) {
-                    m_RigidBody.AddForce(desiredMove * SlopeMultiplier(), ForceMode.Impulse);
+                desiredMove.x = desiredMove.x*movementSettings.CurrentTargetSpeed;
+                desiredMove.z = desiredMove.z*movementSettings.CurrentTargetSpeed;
+                desiredMove.y = desiredMove.y*movementSettings.CurrentTargetSpeed;
+                if (m_RigidBody.velocity.sqrMagnitude <
+                    (movementSettings.CurrentTargetSpeed*movementSettings.CurrentTargetSpeed))
+                {
+                    m_RigidBody.AddForce(desiredMove*SlopeMultiplier(), ForceMode.Impulse);
                 }
             }
 
-            if(m_IsGrounded) {
+            if (m_IsGrounded)
+            {
                 m_RigidBody.drag = 5f;
 
-                if(m_Jump) {
+                if (m_Jump)
+                {
                     m_RigidBody.drag = 0f;
                     m_RigidBody.velocity = new Vector3(m_RigidBody.velocity.x, 0f, m_RigidBody.velocity.z);
                     m_RigidBody.AddForce(new Vector3(0f, movementSettings.JumpForce, 0f), ForceMode.Impulse);
                     m_Jumping = true;
                 }
 
-                if(!m_Jumping && Mathf.Abs(input.x) < float.Epsilon && Mathf.Abs(input.y) < float.Epsilon && m_RigidBody.velocity.magnitude < 1f) {
+                if (!m_Jumping && Mathf.Abs(input.x) < float.Epsilon && Mathf.Abs(input.y) < float.Epsilon && m_RigidBody.velocity.magnitude < 1f)
+                {
                     m_RigidBody.Sleep();
                 }
-            } else {
+            }
+            else
+            {
                 m_RigidBody.drag = 0f;
-                if(m_PreviouslyGrounded && !m_Jumping) {
+                if (m_PreviouslyGrounded && !m_Jumping)
+                {
                     StickToGroundHelper();
                 }
             }
@@ -160,65 +186,77 @@ namespace UnityStandardAssets.Characters.FirstPerson {
         }
 
 
-        private float SlopeMultiplier() {
+        private float SlopeMultiplier()
+        {
             float angle = Vector3.Angle(m_GroundContactNormal, Vector3.up);
             return movementSettings.SlopeCurveModifier.Evaluate(angle);
         }
 
 
-        private void StickToGroundHelper() {
+        private void StickToGroundHelper()
+        {
             RaycastHit hitInfo;
-            if(Physics.SphereCast(transform.position, m_Capsule.radius * (1.0f - advancedSettings.shellOffset), Vector3.down, out hitInfo,
-                                   ((m_Capsule.height / 2f) - m_Capsule.radius) +
-                                   advancedSettings.stickToGroundHelperDistance, Physics.AllLayers, QueryTriggerInteraction.Ignore)) {
-                if(Mathf.Abs(Vector3.Angle(hitInfo.normal, Vector3.up)) < 85f) {
+            if (Physics.SphereCast(transform.position, m_Capsule.radius * (1.0f - advancedSettings.shellOffset), Vector3.down, out hitInfo,
+                                   ((m_Capsule.height/2f) - m_Capsule.radius) +
+                                   advancedSettings.stickToGroundHelperDistance, Physics.AllLayers, QueryTriggerInteraction.Ignore))
+            {
+                if (Mathf.Abs(Vector3.Angle(hitInfo.normal, Vector3.up)) < 85f)
+                {
                     m_RigidBody.velocity = Vector3.ProjectOnPlane(m_RigidBody.velocity, hitInfo.normal);
                 }
             }
         }
 
 
-        private Vector2 GetInput() {
-
-            Vector2 input = new Vector2 {
-                x = Input.GetAxis("Horizontal"),
-                y = Input.GetAxis("Vertical")
-            };
-            movementSettings.UpdateDesiredTargetSpeed(input);
+        private Vector2 GetInput()
+        {
+            
+            Vector2 input = new Vector2
+                {
+                    x = Input.GetAxis("Horizontal"),
+                    y = Input.GetAxis("Vertical")
+                };
+			movementSettings.UpdateDesiredTargetSpeed(input);
             return input;
         }
 
 
-        private void RotateView() {
+        private void RotateView()
+        {
             //avoids the mouse looking if the game is effectively paused
-            if(Mathf.Abs(Time.timeScale) < float.Epsilon)
-                return;
+            if (Mathf.Abs(Time.timeScale) < float.Epsilon) return;
 
             // get the rotation before it's changed
             float oldYRotation = transform.eulerAngles.y;
 
-            mouseLook.LookRotation(transform, cam.transform);
+            mouseLook.LookRotation (transform, cam.transform);
 
-            if(m_IsGrounded || advancedSettings.airControl) {
+            if (m_IsGrounded || advancedSettings.airControl)
+            {
                 // Rotate the rigidbody velocity to match the new direction that the character is looking
                 Quaternion velRotation = Quaternion.AngleAxis(transform.eulerAngles.y - oldYRotation, Vector3.up);
-                m_RigidBody.velocity = velRotation * m_RigidBody.velocity;
+                m_RigidBody.velocity = velRotation*m_RigidBody.velocity;
             }
         }
 
         /// sphere cast down just beyond the bottom of the capsule to see if the capsule is colliding round the bottom
-        private void GroundCheck() {
+        private void GroundCheck()
+        {
             m_PreviouslyGrounded = m_IsGrounded;
             RaycastHit hitInfo;
-            if(Physics.SphereCast(transform.position, m_Capsule.radius * (1.0f - advancedSettings.shellOffset), Vector3.down, out hitInfo,
-                                   ((m_Capsule.height / 2f) - m_Capsule.radius) + advancedSettings.groundCheckDistance, Physics.AllLayers, QueryTriggerInteraction.Ignore)) {
+            if (Physics.SphereCast(transform.position, m_Capsule.radius * (1.0f - advancedSettings.shellOffset), Vector3.down, out hitInfo,
+                                   ((m_Capsule.height/2f) - m_Capsule.radius) + advancedSettings.groundCheckDistance, Physics.AllLayers, QueryTriggerInteraction.Ignore))
+            {
                 m_IsGrounded = true;
                 m_GroundContactNormal = hitInfo.normal;
-            } else {
+            }
+            else
+            {
                 m_IsGrounded = false;
                 m_GroundContactNormal = Vector3.up;
             }
-            if(!m_PreviouslyGrounded && m_IsGrounded && m_Jumping) {
+            if (!m_PreviouslyGrounded && m_IsGrounded && m_Jumping)
+            {
                 m_Jumping = false;
             }
         }
