+++ Fantasy Project.csproj
@@ -13,11 +13,13 @@
     <TargetFrameworkIdentifier>.NETFramework</TargetFrameworkIdentifier>
     <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
     <TargetFrameworkProfile>Unity Subset v3.5</TargetFrameworkProfile>
-    <CompilerResponseFile></CompilerResponseFile>
+    <CompilerResponseFile>
+    </CompilerResponseFile>
     <UnityProjectType>Game:1</UnityProjectType>
     <UnityBuildTarget>StandaloneWindows:5</UnityBuildTarget>
     <UnityVersion>5.5.0f3</UnityVersion>
-    <RootNamespace></RootNamespace>
+    <RootNamespace>
+    </RootNamespace>
     <LangVersion Condition=" '$(VisualStudioVersion)' != '10.0' ">4</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
@@ -52,6 +54,9 @@
     <Reference Include="UnityEngine">
       <HintPath>Library\UnityAssemblies\UnityEngine.dll</HintPath>
     </Reference>
+    <Reference Include="UnityEngine.Advertisements">
+      <HintPath>Library\UnityAssemblies\UnityEngine.Advertisements.dll</HintPath>
+    </Reference>
     <Reference Include="UnityEngine.UI">
       <HintPath>Library\UnityAssemblies\UnityEngine.UI.dll</HintPath>
     </Reference>
@@ -66,6 +71,9 @@
     </Reference>
     <Reference Include="UnityEngine.HoloLens">
       <HintPath>Library\UnityAssemblies\UnityEngine.HoloLens.dll</HintPath>
+    </Reference>
+    <Reference Include="UnityEngine.Purchasing">
+      <HintPath>Library\UnityAssemblies\UnityEngine.Purchasing.dll</HintPath>
     </Reference>
     <Reference Include="UnityEngine.VR">
       <HintPath>Library\UnityAssemblies\UnityEngine.VR.dll</HintPath>
@@ -138,12 +146,14 @@
     <Compile Include="Assets\Scripts\Procedural Terrain\Data\NoiseData.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\Data\TerrainData.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\Data\TextureData.cs" />
-    <Compile Include="Assets\Scripts\Procedural Terrain\Data\Updatable Data.cs" />
+    <Compile Include="Assets\Scripts\Procedural Terrain\Data\Updateable Data.cs" />
+    <Compile Include="Assets\Scripts\Procedural Terrain\DataStructs.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\EndlessTerrain.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\FalloffGenerator.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\HideOnPlay.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\MapDisplay.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\MapGenerator.cs" />
+    <Compile Include="Assets\Scripts\Procedural Terrain\MeshData.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\MeshGenerator.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\ProceduralNoise.cs" />
     <Compile Include="Assets\Scripts\Procedural Terrain\TextureGenerator.cs" />
